// citpwawcl/src/app/models/dati-impianto.ts
export class DatiImpianto {
    codiceImpianto?: string;
    stato?: string;
    dataVar?: Date;
    motivazione?: string;
    tipoImpianto?: string;
    flgApparevvUiExt?: boolean;
    flgContabilizzazione?: boolean;
    stradario?: boolean;
    indirizzoSitad?: string;
    indirizzoNonTrovato?: string;
    comune?: string;
    civico?: string;
    siglaProv?: string;
    provincia?: string;
    pod?: string;
    pdr?: string;
    flgNoPdr?: boolean;
    flgVisuProprietario?: boolean;
    coordX?: number;
    coordY?: number;
    istatComune?: string;
    flgMedioImpiantoCivile?: boolean;
}